Create matmul_datadecompos.c
This Code is designed to study the various data decomposition strategies in a square matrix of order N. The algorithm is timed so as to display the time taken to execute the various levels of decomposition can be displayed. There are 3 kinds of data decomposition undertaken, the row major distribution, column major distribution and the row-col block distribution.